Declaration of delete handler was moved into separate file
