+++ src/proxy.hpp
@@ -108,21 +108,6 @@ public:
 
 	bool initialize(const rapidjson::Value &config);
 
-	struct req_delete
-		: public ioremap::thevoid::simple_request_stream<proxy>
-		, public std::enable_shared_from_this<req_delete>
-	{
-		void on_request(const ioremap::thevoid::http_request &req, const boost::asio::const_buffer &buffer);
-		void on_lookup(const ioremap::elliptics::sync_lookup_result &slr, const ioremap::elliptics::error_info &error);
-		void on_finished(const ioremap::elliptics::sync_remove_result &srr, const ioremap::elliptics::error_info &error);
-
-	private:
-		std::string url_str;
-		ioremap::elliptics::key key;
-		boost::optional<ioremap::elliptics::session> session;
-		size_t total_size;
-	};
-
 	struct req_ping
 		: public ioremap::thevoid::simple_request_stream<proxy>
 		, public std::enable_shared_from_this<req_ping>
